Add builder for test server
I need to use this for benchmarking so would like
to organize this a bit better upfront.
