+++ tests/it/cache.rs
@@ -1,7 +1,7 @@
 use serde_json::{Value, json};
 use test_utils::{StatusCode, TestClient, TestResult};
 
-use crate::TestContext;
+use crate::{TestContext, server::start_server};
 
 async fn cache_set(client: &TestClient, key: &str, expire_in: u64, value: &str) -> TestResult<()> {
     client
@@ -34,7 +34,7 @@ async fn test_cache_set_and_get() -> TestResult {
         client,
         handle: _handle,
         ..
-    } = super::start_server().await;
+    } = start_server().await;
 
     cache_set(&client, "test-key-1", 60000, "test-value-123").await?;
 
@@ -53,7 +53,7 @@ async fn test_cache_set_get_and_delete() -> TestResult {
         client,
         handle: _handle,
         ..
-    } = super::start_server().await;
+    } = start_server().await;
 
     cache_set(&client, "test-key-2", 30000, "another-value").await?;
